app/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.8.0
APScheduler==3.11.0
arrow==1.3.0
async-timeout==4.0.3
asyncer==0.0.8
attrs==25.1.0
Authlib==1.3.1
backoff==2.2.1
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
databricks-sdk==0.44.0
dataclasses-json==0.6.7
datasets==3.2.0
Deprecated==1.2.18
diff-match-patch==20230430
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dspy==2.6.3
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==25.9.2
fastapi==0.115.8
fastapi-sso==0.16.0
fastembed==0.4.0
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.68.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
griffe==0.36.9
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
guardrails-ai==0.6.3
guardrails-api-client==0.4.0a1
guardrails_hub_types==0.0.4
gunicorn==22.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.36.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langgraph==0.2.70
langgraph-checkpoint==2.0.12
langgraph-sdk==0.1.51
langsmith==0.3.7
lark==1.1.9
litellm==1.61.3
loguru==0.7.3
lxml==4.9.4
magicattr==0.1.6
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
mdurl==0.1.2
mlflow==2.20.2
mlflow-skinny==2.20.2
mmh3==4.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.19.2
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==4.2.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydash==7.0.7
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.9.4
rpds-py==0.22.3
rq==2.1.0
rsa==4.9
rstr==3.2.2
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
shellingham==1.5.4
simpleeval==1.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.38
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlparse==0.5.3
starlette==0.45.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.29.0
uvloop==0.21.0
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0